refactor: ProductCollectionViewCell 프로토콜 기본구현
