+++ OpenMarket/OpenMarket/Protocol/ProductCollectionViewCell.swift
@@ -17,3 +17,24 @@ protocol ProductCollectionViewCell: UICollectionViewCell {
   func setCellImage(image: UIImage?)
   func setCellData(product: Product)
 }
+
+extension ProductCollectionViewCell {
+  func setCellImage(image: UIImage?) {
+    productImageView.image = image
+  }
+  
+  func setCellData(product: Product) {
+    productNameLabel.text = product.name
+    productFixedPriceLabel.attributedText = product.formattedFixedPrice.strikeThrough(strikeTarget: product.formattedFixedPrice)
+    productBargainPriceLabel.text = product.formattedBargainPrice
+    if product.formattedFixedPrice == product.formattedBargainPrice {
+      productFixedPriceLabel.isHidden = true
+    }
+    if product.formattedStock == "품절" {
+      productStockLabel.textColor = .orange
+    } else {
+      productStockLabel.textColor = .darkGray
+    }
+    productStockLabel.text = product.formattedStock
+  }
+}
