+++ OpenMarket/OpenMarket/View/ProductCollectionViewGridCell.swift
@@ -20,25 +20,6 @@ class ProductCollectionViewGridCell: UICollectionViewCell, ReuseIdentifying, Pro
     configureBorderStyle()
   }
   
-  func setCellImage(image: UIImage?) {
-    productImageView.image = image
-  }
-  
-  func setCellData(product: Product) {
-    productNameLabel.text = product.name
-    productFixedPriceLabel.attributedText = product.formattedFixedPrice.strikeThrough(strikeTarget: product.formattedFixedPrice)
-    productBargainPriceLabel.text = product.formattedBargainPrice
-    if product.formattedFixedPrice == product.formattedBargainPrice {
-      productFixedPriceLabel.isHidden = true
-    }
-    if product.formattedStock == "품절" {
-      productStockLabel.textColor = .orange
-    } else {
-      productStockLabel.textColor = .darkGray
-    }
-    productStockLabel.text = product.formattedStock
-  }
-  
   private func configureBorderStyle() {
     contentView.layer.borderWidth = 1
     contentView.layer.borderColor = UIColor.gray.cgColor
