--- OpenMarket/OpenMarket/View/ProductImageRegistrationCell.swift
@@ -1,16 +1,0 @@
-//
-//  ProductImageRegistrationCell.swift
-//  OpenMarket
-//
-//  Created by 박병호 on 2022/01/19.
-//
-
-import UIKit
-
-class ProductImageRegistrationCell: UICollectionViewCell, ReuseIdentifying {
-  @IBOutlet weak var imageView: UIImageView!
-  
-  override func awakeFromNib() {
-    
-  }
-}
